svc/pyproject.toml
[project]
name = "glazing-control-service"
version = "0.1.0"
description = "Electrochromic glazing control service for Trailer Lab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn==0.32.1",
    "pydantic==2.9.2",
    "httpx==0.27.2",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "pyserial>=3.5",
    "numpy==2.4.2",
    "scipy==1.17.0",
    "colour-science==0.4.7",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
